Change pwm_set_dc() signature to accept a float number instead of uint16_t

+++ tp-final/Drivers/API/inc/API_pwm.h
@@ -4,4 +4,4 @@
 
 void pwm_init();
 
-void pwm_set_dc(const uint16_t duty_cycle);
+void pwm_set_dc(const float duty_cycle);
